LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/back-forward-out-of-bounds-expected.txt



FAIL back() and forward() out of bounds assert_equals: expected 1 but got 0
